+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.4 (April 28, 2020)
+
+* Add new props to customization buttons (VisibilityButtonProps, CopyToClipboardButtonProps, GenerateButtonProps);
+* Add new props to customization icons (AutorenewIconProps, VisibilityIconProps, VisibilityOffIconProps, FileCopyOutlinedIconProps)
+
 ## 1.0.3 (April 27, 2020)
 
 * Fix copy to clipboard bug